Convert Sidebar to ConditionalTagCheck
Simplify and generalize class since it had nothing to do directly with
the sidebar. Now it can be used for other purposes as well.
