+++ functions.php
@@ -10,16 +10,16 @@
  * @link https://github.com/roots/sage/pull/1042
  */
 $sage_includes = [
-  'lib/utils.php',           // Utility functions
-  'lib/init.php',            // Initial theme setup and constants
-  'lib/wrapper.php',         // Theme wrapper class
-  'lib/sidebar.php',         // Sidebar class
-  'lib/config.php',          // Configuration
-  'lib/assets.php',          // Scripts and stylesheets
-  'lib/titles.php',          // Page titles
-  'lib/nav.php',             // Custom nav modifications
-  'lib/gallery.php',         // Custom [gallery] modifications
-  'lib/extras.php',          // Custom functions
+  'lib/utils.php',                 // Utility functions
+  'lib/init.php',                  // Initial theme setup and constants
+  'lib/wrapper.php',               // Theme wrapper class
+  'lib/conditional-tag-check.php', // ConditionalTagCheck class
+  'lib/config.php',                // Configuration
+  'lib/assets.php',                // Scripts and stylesheets
+  'lib/titles.php',                // Page titles
+  'lib/nav.php',                   // Custom nav modifications
+  'lib/gallery.php',               // Custom [gallery] modifications
+  'lib/extras.php',                // Custom functions
 ];
 
 foreach ($sage_includes as $file) {
